+++ content-none.php
@@ -7,6 +7,7 @@
  * @package Primer
  * @since   1.0.0
  */
+
 ?>
 
 <section class="no-results not-found">
@@ -21,7 +22,8 @@
 
 		<?php if ( is_home() && current_user_can( 'publish_posts' ) ) : ?>
 
-			<p><?php
+			<p>
+			<?php
 
 			printf(
 				esc_html_x( 'Ready to publish your first post? %s.', 'link to write a new post', 'primer' ),
@@ -32,7 +34,8 @@
 				)
 			);
 
-			?></p>
+			?>
+			</p>
 
 		<?php elseif ( is_search() ) : ?>