+++ inc/customizer/fonts.php
@@ -280,7 +280,7 @@ class Primer_Customizer_Fonts {
 	 */
 	public function sanitize_font( $font ) {
 
-		return in_array( $font, $this->fonts ) ? $font : $this->fonts[0];
+		return in_array( $font, $this->fonts, true ) ? $font : $this->fonts[0];
 
 	}