+++ .dev/docs/sphinx/src/changelog.md
@@ -1,10 +1,19 @@
 ## Changelog
 
+### 1.5.0 ###
+* New: Full compatibility with the WooCommerce plugin @props [evanherman](https://github.com/EvanHerman), [fjarrett](https://github.com/fjarrett)
+* New: Add Front Page Title setting to the Static Front Page section of the Customizer @props [evanherman](https://github.com/EvanHerman)
+* New: Added Footer Copyright Text setting to the Site Identity section of the Customizer @props [evanherman](https://github.com/EvanHerman)
+* New: Added Author Credit toggle to the Site Identity section of the Customizer @props [evanherman](https://github.com/EvanHerman)
+* New: [Developer documentation!](https://godaddy.github.io/wp-primer-theme/) @props [evanherman](https://github.com/EvanHerman)
+* Tweak: Added `nofollow` attribute to theme author link in the footer @props [evanherman](https://github.com/EvanHerman)
+* Tweak: Use core post navigation functions @props [evanherman](https://github.com/EvanHerman)
+* Fix: Display glitch when last main menu item is a submenu @props [evanherman](https://github.com/EvanHerman)
+
 ### 1.4.2 ###
 * Added word-wrap where text was overflowing @props [evanherman](https://github.com/EvanHerman)
 * Updated translations @props [jonathanbardo](https://github.com/jonathanbardo)
 * Added font-family support for WooCommerce buttons @props [fjarret](https://github.com/fjarrett)
-* Introduced 'Author Credit' toggle in the customizer @props [evanherman](https://github.com/EvanHerman)
 
 ### 1.4.1 ###
 * Fixed hero widget font sizes @props [jonathanbardo](https://github.com/jonathanbardo)